+++ src/pages/FlowRun/Details-Tile.vue
@@ -44,6 +44,12 @@ export default {
         this.flowRun.flow.parameters.length > 0
       )
     },
+    hasRunConfig() {
+      return (
+        this.flowRun?.run_config &&
+        Object.keys(this.flowRun.run_config).length > 0
+      )
+    },
     flowRunParams() {
       const flowParams = this.flowRun?.flow?.parameters.reduce(
         (accum, currentParam) => {
@@ -82,7 +88,10 @@ export default {
       <!-- <v-icon>{{ flowRun.state }}</v-icon> -->
     </v-system-bar>
 
-    <CardTitle v-if="hasParameters || hasContext" icon="pi-flow-run">
+    <CardTitle
+      v-if="hasParameters || hasContext || hasRunConfig"
+      icon="pi-flow-run"
+    >
       <v-row slot="title" no-gutters class="d-flex align-center">
         <v-col cols="8">
           <div class="text-truncate pb-1">
@@ -158,6 +167,29 @@ export default {
         >
           Context
           <v-icon small>list</v-icon>
+        </v-btn>
+
+        <v-btn
+          v-if="hasRunConfig"
+          depressed
+          small
+          tile
+          icon
+          class="button-transition w-100 d-flex justify-end"
+          :color="tab == 'run_config' ? 'primary' : ''"
+          :style="{
+            'border-right': `3px solid ${
+              tab == 'run_config'
+                ? 'var(--v-primary-base)'
+                : 'var(--v-appForeground-base)'
+            }`,
+            'box-sizing': 'content-box',
+            'min-width': '100px'
+          }"
+          @click="tab = 'run_config'"
+        >
+          Run Config
+          <v-icon small>notes</v-icon>
         </v-btn>
       </div>
     </CardTitle>
@@ -274,9 +306,7 @@ export default {
                   </v-col>
                 </v-row>
                 <v-row v-if="flowRun.end_time" no-gutters>
-                  <v-col cols="6">
-                    Ended
-                  </v-col>
+                  <v-col cols="6"> Ended </v-col>
                   <v-col cols="6" class="text-right font-weight-bold">
                     <v-tooltip top>
                       <template #activator="{ on }">
@@ -291,9 +321,7 @@ export default {
                   </v-col>
                 </v-row>
                 <v-row v-if="flowRun.start_time" no-gutters>
-                  <v-col cols="6">
-                    Duration
-                  </v-col>
+                  <v-col cols="6"> Duration </v-col>
                   <v-col cols="6" class="text-right font-weight-bold">
                     <!-- Check for isFinished improves duration handling for restarted flows  -->
                     <DurationSpan
@@ -332,6 +360,15 @@ export default {
           class="text-body-2 appForeground rounded-sm pa-5 code-block"
         >
           {{ formatJson(flowRun.context) }}
+        </div>
+      </v-fade-transition>
+
+      <v-fade-transition hide-on-leave>
+        <div
+          v-if="tab === 'run_config'"
+          class="text-body-2 appForeground rounded-sm pa-5 code-block"
+        >
+          {{ formatJson(flowRun.run_config) }}
         </div>
       </v-fade-transition>
     </v-card-text>